[Contract API] Refactor API for better format
