+++ routes/contract.js
@@ -3,13 +3,15 @@ var contract = require('../models/contract.js');
 var contractFunction = require('../models/contractFunction.js');
 var contractEvent = require('../models/contractEvent.js');
 var contractController = require('../controllers/contractController.js');
+var transactionData = require('../models/transactionData.js');
+var eventData = require('../models/eventData.js');
 var sqsHelper = require('../helpers/aws/sqsHelper.js');
 var router = express.Router();
 
 /**
  * GET contract info
  */
-router.get('/v1/info', function (req, res, next) {
+router.get('/info', function (req, res, next) {
     var contractId = req.query.contractId;
     var contractInfo = null;
     var contractEventInfo = null;
@@ -36,7 +38,7 @@ router.get('/v1/info', function (req, res, next) {
 /**
  * GET load contract event
  */
-router.get('/v1/event', function(req, res, next) {
+router.get('/events', function(req, res, next) {
 	if (req.query.contractEventId === undefined) {
 		res.json({'error': {'code': 214, 'message': 'contractEventId is null'}});
 	}
@@ -56,7 +58,7 @@ router.get('/v1/event', function(req, res, next) {
 /**
  * GET load contract function
  */
-router.get('/v1/function', function(req, res, next) {
+router.get('/functions', function(req, res, next) {
 	if (req.query.contractFunctionId === undefined) {
 		res.json({'error': {'code': 215, 'message': 'contractFunctionId is null'}});
 	}
@@ -70,6 +72,76 @@ router.get('/v1/function', function(req, res, next) {
 	}).catch(function(err) {
 		console.log(err.message, err.stack);
 		res.json({'error': {'code': 213, 'message': 'error on load data'}});
+	});
+});
+
+/**
+ * POST use contract function
+ */
+router.post('/tx', function(req, res, next) {
+	if (req.body.contractId === undefined) {
+		res.json({'error': {'code': 201, 'message': 'contractId is null'}});
+	}
+	if (req.body.contractFunctionId === undefined) {
+		res.json({'error': {'code': 202, 'message': 'contractFunctionId is null'}});
+	}
+	if (req.body.data === undefined) {
+		res.json({'error': {'code': 203, 'message': 'data is null'}});
+	}
+	var contractId = req.body.contractId;
+	var contractFunctionId = req.body.contractFunctionId;
+	var data = req.body.data;
+	transactionData.create({"data": JSON.stringify(data)}).then(function (txHash) {
+		var message = {
+			"txHash": txHash,
+			"contractId": contractId,
+			"contractFunctionId": contractFunctionId,
+			"data": data
+		};
+		sqsHelper.send(JSON.stringify(message), 
+					process.env.AWS_TRANSACTION_QUEUE_URL, 10, 
+					'transaction');
+		res.json({'data': {'txHash': txHash}});
+	}).catch(function(err) {
+		console.log(err.message, err.stack);
+		res.json({'error': {'code': 204, 'message': 'error on send transaction'}});
+	});
+});
+
+/**
+ * GET function transaction status
+ */
+router.get('/tx', function(req, res, next){
+	if (req.query.txHash === undefined) {
+		res.json({'error': {'code': 211, 'message': 'txHash is null'}});
+	}
+	var txHash = req.query.txHash;
+	transactionData.read(txHash).then(function(transactionDataResult) {
+		if (transactionDataResult.rowCount > 0) {
+			res.json({'data': transactionDataResult.rows[0]});
+		} else {
+			res.json({'error': {'code': 212, 'message': 'empty data'}});
+		}
+	}).catch(function(err) {
+		console.log(err.message, err.stack);
+		res.json({'error': {'code': 213, 'message': 'error on load data'}});
+	});
+});
+
+/**
+ * GET load event data 
+ */
+router.get('/event/data', function(req, res, next) {
+	if (req.query.txHash === undefined) {
+		res.json({'error': {'code': 211, 'message': 'txHash is null'}});
+	}
+	var txHash = req.query.txHash;
+	eventData.readByTxHash(txHash).then(function(result) {
+		if (result.rowCount > 0) {
+			res.json({'data': result.rows});
+		} else {
+			res.json({'error': {'code': 212, 'message': 'empty data'}});
+		}
 	});
 });
